Finish refactor of loading / saving

+++ src/openrct2/scenario/Scenario.cpp
@@ -65,7 +65,7 @@ bool gFirstTimeSaving = true;
 uint16_t gSavedAge;
 uint32_t gLastAutoSaveUpdate = 0;
 
-uint64_t gScenarioTicks;
+uint32_t gScenarioTicks;
 uint32_t gScenarioSrand0;
 uint32_t gScenarioSrand1;
